+++ src/components/Transfer/index.ts
@@ -1,0 +1,7 @@
+export { Transfer } from './Transfer'
+export type {
+  TransferProps,
+  TransferItem,
+  TransferTargetEntry,
+  TransferGroup,
+} from './Transfer'